refactor: extract drain domain (write side) into src/drain/
