+++ test/persist.test.ts
@@ -1,11 +1,10 @@
 import { describe, it, expect, afterEach } from "vitest";
-import { mkdtempSync, rmSync, readdirSync } from "node:fs";
+import { mkdtempSync, rmSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { createCollector } from "../src/capture/collector";
-import { drainToDisk } from "../src/persist/drain-to-disk";
+import { drainToDisk } from "../src/drain/to-disk";
 import { readCatalog } from "../src/persist/read-catalog";
-import { endpointSlug } from "../src/persist/slug";
 import type { CapturedRequest, CapturedResponse } from "../src/types";
 
 let dir: string;
@@ -20,19 +19,7 @@ const seed = (name: string, body: unknown) => {
   return c;
 };
 
-describe("persist", () => {
-  it("slugs an endpoint deterministically", () => {
-    expect(endpointSlug("GET", "/users/:id")).toBe("get-users-id");
-  });
-
-  it("drains uniquely-named files (EC7: no clobber across processes)", () => {
-    dir = mkdtempSync(join(tmpdir(), "owl-"));
-    const files = drainToDisk(seed("(200) ok", { id: 1 }), dir);
-    expect(files).toHaveLength(1);
-    const [name] = readdirSync(dir);
-    expect(name).toMatch(/^get-users-id\..+\.json$/); // <slug>.<unique>.json
-  });
-
+describe("catalog (merge)", () => {
   it("merges Examples by endpoint key across multiple drain files (EC7)", () => {
     dir = mkdtempSync(join(tmpdir(), "owl-"));
     // simulate two separate test processes documenting the SAME endpoint
